update eventual consistency guarantees

+++ consensus/entry.go
@@ -2,10 +2,10 @@ package consensus
 
 type Entry struct {
 	key, value []byte
-	version int64
+	version int
 }
 
-func NewEntry(key, value []byte, version int64) Entry {
+func NewEntry(key, value []byte, version int) Entry {
 	return Entry{
 		key: key,
 		value: value,
